Restructured examples/, adjusted readme
