+++ examples/sil_example.py
@@ -10,8 +10,8 @@ This is example experimental and documentation is still in progress.
 
 import mosaik  # type: ignore
 
-from examples._data import load_carbon_data, load_solar_data
-from examples.cosim_example.cosim_example import (
+from _data import load_carbon_data, load_solar_data
+from cosim_example import (
     COSIM_CONFIG,
     SIM_START,
     STORAGE,
@@ -28,7 +28,7 @@ from vessim.cosim._util import disable_mosaik_warnings
 COSIM_SIL_CONFIG = {
     **COSIM_CONFIG,
     "SilInterface": {
-        "python": "vessim.sil_example.sil_interface:SilInterfaceSim",
+        "python": "vessim.sil.sil_interface:SilInterfaceSim",
     },
 }
 RT_FACTOR = 1/60  # 1 wall-clock second ^= 60 sim seconds
