+++ examples/util/cacu.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 from vessim.sil.http_client import HttpClient
 from vessim.sil.loop_thread import LoopThread
-from examples.cosim_example.cacu import cacu_scenario
+from simulated_cacu import cacu_scenario
 
 
 class RemoteBattery:
