+++ examples/util/example_node/rpi/init.sh
