+++ examples/util/example_node/virtual_node/linear_power_model.py
