+++ examples/util/tf_aws_node/main.tf
