+++ examples/util/tf_aws_node/outputs.tf
