+++ examples/util/tf_aws_node/startup.sh
