+++ examples/util/tf_aws_node/variables.tf
