+++ examples/util/tf_gcp_node/main.tf
