+++ examples/util/tf_gcp_node/outputs.tf
