+++ examples/util/tf_gcp_node/startup.sh
