+++ examples/util/tf_gcp_node/variables.tf
